refactor: shared sement들 분리
