+++ src/app/ui/globalLayout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom"
-import Header from "../../components/Header"
-import Footer from "../../components/Footer"
+import Header from "../../shared/ui/Header"
+import Footer from "../../shared/ui/Footer"
 
 const GlobalLayout = () => {
   return (
