+++ src/entities/comments/api/deleteComment.ts
@@ -1,4 +1,4 @@
-import fetchClient from "../../../shared/utils/fetchClient"
+import fetchClient from "../../../shared/api/fetchClient"
 
 export const deleteCommentApi = async (id: number): Promise<void> => {
   return fetchClient<void>(`/comments/${id}`, {
