+++ src/entities/posts/api/deletePost.ts
@@ -1,4 +1,4 @@
-import fetchClient from "../../../shared/utils/fetchClient"
+import fetchClient from "../../../shared/api/fetchClient"
 
 export const deletePostApi = async (id: number): Promise<void> => {
   return fetchClient<void>(`/posts/${id}`, {
