+++ src/entities/tags/api/fetchTags.ts
@@ -1,4 +1,4 @@
-import fetchClient from "../../../shared/utils/fetchClient"
+import fetchClient from "../../../shared/api/fetchClient"
 import { Tag } from "./types"
 
 export const fetchTagsApi = async (): Promise<Tag[]> => {
