+++ src/shared/ui/Footer.tsx
